# frontend/11心理健康聊天机器人 API.openapi.json
{
  "openapi": "3.0.1",
  "info": {
    "title": "11心理健康聊天机器人 API",
    "description": "",
    "version": "1.0.0"
  },
...
  "paths": {
...
    "/api/auth/register": {
      "post": {
        "summary": "Register",
        "deprecated": false,
        "description": "用户注册",
        "operationId": "register_api_auth_register_post",
        "tags": [
          "Authentication"
        ],
        "parameters": [],
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/RegisterRequest"
              }
            }
          }
        },
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/AuthResponse"
                }
              }
            },
            "headers": {}
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            },
            "headers": {}
          }
        },
        "security": [
          {
            "bearer": []
          }
        ]
      }
    },
    "/api/auth/login": {
      "post": {
        "summary": "Login",
        "deprecated": false,
        "description": "用户登录",
        "operationId": "login_api_auth_login_post",
        "tags": [
          "Authentication"
        ],
        "parameters": [],
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/LoginRequest"
              },
              "example": {
                "email": "user@example.com",
                "password": "new123456"
              }
            }
          }
        },
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/AuthResponse"
                },
                "example": {
                  "access_token": "eyJhbGciOiJIUzI1NiIsInR5cCI6IkpXVCJ9.eyJzdWIiOiI2ZmUxZjFhZS1lNWI4LTRmMGQtODJlZi00ODY5YTIzNmQ1ZGIiLCJleHAiOjE3NTM1Njk1MzZ9.0t33sTz34YGmfdkHplqIhBNNLPeZ7EXVkVLwsyYNe9A",
                  "refresh_token": "eyJhbGciOiJIUzI1NiIsInR5cCI6IkpXVCJ9.eyJzdWIiOiI2ZmUxZjFhZS1lNWI4LTRmMGQtODJlZi00ODY5YTIzNmQ1ZGIiLCJleHAiOjE3NTQxNzI1MzZ9.xfit8wAW_nohfQdzQ8DKw-ORZqraJ_7bpVls6rttf_U",
                  "token_type": "bearer",
                  "user_id": "6fe1f1ae-e5b8-4f0d-82ef-4869a236d5db",
                  "email": "user@example.com",
                  "username": "张三"
                }
              }
            },
            "headers": {}
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            },
            "headers": {}
          }
        },
        "security": [
          {
            "bearer": []
          }
        ]
      }
    },
    "/api/auth/change-password": {
      "post": {
        "summary": "修改密码",
        "deprecated": false,
        "description": "",
        "tags": [],
        "parameters": [
          {
            "name": "Content-Type",
            "in": "header",
            "description": "",
            "required": false,
            "example": "application/json",
            "schema": {
              "type": "string"
            }
          },
          {
            "name": "Authorization",
            "in": "header",
            "description": "",
            "required": false,
            "example": "Bearer {{access_token}}",
            "schema": {
              "type": "string"
            }
          }
        ],
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "type": "object",
                "properties": {}
              },
              "example": {
                "current_password": "new123456",
                "new_password": "password123"
              }
            }
          }
        },
        "responses": {
          "200": {
            "description": "",
            "content": {
              "application/json": {
                "schema": {
                  "type": "object",
                  "properties": {}
                }
              }
            },
            "headers": {}
          }
        },
        "security": [
          {
            "bearer": []
          }
        ]
      }
    },
    "/api/auth/refresh": {
      "post": {
        "summary": "Refresh Token",
        "deprecated": false,
        "description": "刷新访问令牌",
        "operationId": "refresh_token_api_auth_refresh_post",
        "tags": [
          "Authentication"
        ],
        "parameters": [
          {
            "name": "refresh_token",
            "in": "query",
            "description": "",
            "required": true,
            "example": "eyJhbGciOiJIUzI1NiIsInR5cCI6IkpXVCJ9.eyJzdWIiOiI2ZmUxZjFhZS1lNWI4LTRmMGQtODJlZi00ODY5YTIzNmQ1ZGIiLCJleHAiOjE3NTQxNTAyOTB9.gtUd_J0819r-dvUR7_yirD83HKBRlI5MYRto3rNaifY",
            "schema": {
              "type": "string",
              "title": "Refresh Token"
            }
          }
        ],
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "type": "object",
                "properties": {}
              },
              "examples": {}
            }
          }
        },
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {}
              }
            },
            "headers": {}
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            },
            "headers": {}
          }
        },
        "security": [
          {
            "bearer": []
          }
        ]
      }
    },
    "/api/auth/validate": {
      "post": {
        "summary": "Validate Token",
        "deprecated": false,
        "description": "验证令牌",
        "operationId": "validate_token_api_auth_validate_post",
        "tags": [
          "Authentication"
        ],
        "parameters": [],
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {}
              }
            },
            "headers": {}
          }
        },
        "security": [
          {
            "bearer": []
          }
        ]
      }
    },
    "/api/auth/logout": {
      "post": {
        "summary": "Logout",
        "deprecated": false,
        "description": "用户登出",
        "operationId": "logout_api_auth_logout_post",
        "tags": [
          "Authentication"
        ],
        "parameters": [],
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {}
              }
            },
            "headers": {}
          }
        },
        "security": [
          {
            "bearer": []
          }
        ]
      }
    },
    "/api/auth/me": {
      "get": {
        "summary": "Get Current User Info",
        "deprecated": false,
        "description": "获取当前用户信息",
        "operationId": "get_current_user_info_api_auth_me_get",
        "tags": [
          "Authentication"
        ],
        "parameters": [
          {
            "name": "Authorization",
            "in": "header",
            "description": "",
            "required": true,
            "example": "Bearer {{access_token}}",
            "schema": {
              "type": "string"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/UserInfo"
                }
              }
            },
            "headers": {}
          }
        },
        "security": [
          {
            "bearer": []
          }
        ]
      }
    },
...
  },
  "components": {
    "schemas": {
...
      "AuthResponse": {
        "properties": {
          "access_token": {
            "type": "string",
            "title": "Access Token"
          },
          "refresh_token": {
            "type": "string",
            "title": "Refresh Token"
          },
          "token_type": {
            "type": "string",
            "title": "Token Type",
            "default": "bearer"
          },
          "user_id": {
            "type": "string",
            "title": "User Id"
          },
          "email": {
            "type": "string",
            "title": "Email"
          },
          "username": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Username"
          }
        },
        "type": "object",
        "required": [
          "access_token",
          "refresh_token",
          "user_id",
          "email"
        ],
        "title": "AuthResponse"
      },
...
      "UserInfo": {
        "properties": {
          "user_id": {
            "type": "string",
            "title": "User Id"
          },
          "email": {
            "type": "string",
            "title": "Email"
          },
          "username": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Username"
          },
          "is_active": {
            "type": "boolean",
            "title": "Is Active"
          },
          "created_at": {
            "type": "string",
            "title": "Created At"
          }
        },
        "type": "object",
        "required": [
          "user_id",
          "email",
          "is_active",
          "created_at"
        ],
        "title": "UserInfo"
      },
      "HTTPValidationError": {
        "properties": {
          "detail": {
            "items": {
              "$ref": "#/components/schemas/ValidationError"
            },
            "type": "array",
            "title": "Detail"
          }
        },
        "type": "object",
        "title": "HTTPValidationError"
      },
...
      "LoginRequest": {
        "properties": {
          "email": {
            "type": "string",
            "format": "email",
            "title": "Email",
            "description": "用户邮箱"
          },
          "password": {
            "type": "string",
            "minLength": 6,
            "title": "Password",
            "description": "用户密码"
          }
        },
        "type": "object",
        "required": [
          "email",
          "password"
        ],
        "title": "LoginRequest"
      },
...
      "RegisterRequest": {
        "properties": {
          "username": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Username",
            "description": "用户名"
          },
          "email": {
            "type": "string",
            "format": "email",
            "title": "Email",
            "description": "用户邮箱"
          },
          "password": {
            "type": "string",
            "minLength": 6,
            "title": "Password",
            "description": "用户密码"
          }
        },
        "type": "object",
        "required": [
          "email",
          "password"
        ],
        "title": "RegisterRequest"
      },
...
      "ValidationError": {
        "properties": {
          "loc": {
            "items": {
              "anyOf": [
                {
                  "type": "string"
                },
                {
                  "type": "integer"
                }
              ]
            },
            "type": "array",
            "title": "Location"
          },
          "msg": {
            "type": "string",
            "title": "Message"
          },
          "type": {
            "type": "string",
            "title": "Error Type"
          }
        },
        "type": "object",
        "required": [
          "loc",
          "msg",
          "type"
        ],
        "title": "ValidationError"
      },
...
    },
    "securitySchemes": {
      "bearerAuth": {
        "type": "http",
        "scheme": "bearer",
        "bearerFormat": "JWT"
      },
      "HTTPBearer": {
        "type": "http",
        "scheme": "bearer"
      },
      "bearer": {
        "type": "http",
        "scheme": "bearer"
      }
    }
  },
  "servers": [],
  "security": [
    {
      "bearer": []
    }
  ]
}